Added basic parser
Generics needed to be removed from the Expr types, but they don't seem
to have been adding anything anyways.

Fixed a bug around literals in the scanner. Need to revisit that.

Reorganized the "main" code into a `Lox` class; also reorganized error
reporting slightly.

Added cases to `LiteralValue` for false, true, and nil. It's less
useful than it initially seemed; not clear whether it should be removed
in favor of `Any` or strengthened somehow.