Added error handling to interpreter
Corrected parser bugs: treating initial - as an erroneous addition
expression; failing to parse nested parentheses.

The TokenKind+Interpretation file is broken out of Interpreter because
the compiler must see the original definition of the nested type before
trying to compile the extension, and there doesn't seem to be any way
to specify compilation order in the SPM other than filename.