Added variable declaration and use
A variable can be declared with a `var` statement, and given an optional
initializer expression, defaulting to `nil`.

Using a previously-declared name produces the currently stored value.
It is a runtime error to refer to a name that was not previously
declared.

The `Environment` class manages variable state for the `Interpreter`.

Redeclaration is allowed. Added a test file to demonstrate.

`Parser().synchronize()` is now active.