Added requirement to initialize a var before use
Runtime values are now represented by `LoxValue` instead of `Any`. Then
in the `Environment`, `nil` represents a truly missing value instead of
being equivalent to Lox's own `nil`. Trying to use such a missing value
is a runtime error.

The added layer of indirection is also helpful for clarity in the
`Interpreter`.