Classes: decl, instantiation, member get/set
Class bodies declare methods. At runtime a class is an instance of
`LoxClass`. Invoking a class name produces an instance, represented by
a `LoxInstance`. An instance can have member accesses: gets and sets,
where the member is set off from the object by a dot. These accesses
are composable.

Method declarations are handled, but method calls are not yet.