Added .this references inside methods
Rather than following the book and using a scope/environment immediately
outside the method, there is a hidden parameter and argument added to the
method during resolution and interpretation. The `VariableResolver` and
`Interpreter` both add a synthesized `.this` token to the parameter list.

When a method is looked up on an object, the object binds itself to the
`Callable`, producing a new `Callable` with a reference to the
`LoxInstance`. On invocation, the `Callable` prepends that bound object
to the argument list. (It might be good to cache these bound methods.)

Removed a bogus assertion for variable lookup failure at runtime (it is
still a runtime error, but within Lox, not in Swift).