Report use of .this outside methods as an error
Use as value is detected in static analysis; `ControlStatementAnalyzer`
has been changed to `KeywordPlacementAnalyzer` to support this.

Declaration as a new variable (or parameter name) is also disallowed;
this is handled by the parser, but without terminating parsing. Misuses
of `super` as an identifier are also caught at parse time.