Added support for initializers
A method with the special name `init` will be plucked out of a class's
method list when the declaration is interpreted and stored separately.
(This prevents the method from being invoked on an instance.)
When a call expression for the class is interpreted, the initializer will
be invoked with the newly-created instance as the first argument, like
any other method.

Classes are permitted to declare no initializer. In this case, a call
expression will simply return a new instance, unmodified.