Added "getters" to classes (Challenge 2 in Ch. 12)
Classes can now include named executable blocks that have no parameter
list and always produce a value. Like a normal method, the getter is
bound to the instance from which it is accessed. The static analysis
stage enforces that the last statement in the getter's body must be a
return statement.

The `Callable` that represents a getter is marked as being "implicitly
invoked"; when an `Expression.get` produces such a `Callable`, the
interpreter immediately invokes it and returns the result instead of the
`.callable` itself. (This means that there is no way to refer to a
getter, using it as its own value.)