Added inheritance and super expressions
Classes can specify a superclass; it is then legal to call methods from
the superclass on an instance of the subclass. A class can also refer to
its own superclass inside a method with the `super` keyword. This causes
method resolution to start with that superclass.

To facilitate `super` resolution, a new environment/scope is created when
a class declaration is processed, to contain the "super" value.

Changed format of error/warning messages to be closer to clang's.