Add some high-level tests for StringCompiler
The tests are comparing `StringCompiler`'s rendering to Swift's own,
using Swift's own implementation of equality. This is more
straightforward than doing detailed validations of the UTF-8 itself.

Small refactor of `StringCompiler` to accept just the allocate/destroy
function pair that it needs, instead of depending on `MemoryManager`.

Add a test target that shares source files with the loxvm target; it
excludes main.swift and so is a valid dependency for a unit test\
target. This requires tricking SPM into thinking that there are
separate files; for some reason it is an error for two targets to
compile the same files, even if there's no dependency relationship. So
there's now a symlink Sources/loxvm-testable for the Sources/loxvm
directory.