First step for string interning in the Compiler
Lox's hash function is a global function so that it can be
used to check for uniqueness at string compilation time.

Rename `Table` to `HashTable` and make it a class -- especially
in the case of the `strings` table its whole purpose is to be
a single well-known instance. Add a basic unit test for the
`HashTable`.

Update to Swift 5.3 -- assorted syntax modifications required.