Add tests for HashTable
Basic assertions for expansion and deletion, plus high-ish-volume
interning of strings just to try to flush out memory error regression.

The Package includes a testable target for loxvm that does not
include main.swift so that unit tests can be run directly against
it. SPM doesn't allow two targets to share source files for some
reason, so the sources for loxvm-testable are symlinks to loxvm's.