Deduplicate global variable names in constants
(Challenge 1 in Chapter 21)

Rather than creating a new entry in the chunk's constants list for
every global variable operation, we take advantage of the compiler's
string uniquing and look for an existing instance of the string first.
If found, that index is used as the operation's argument.

Added a generated test file to exercise global definition, setting, and
lookup.