Perform global name deduping in Compiler, not Chunk
Rather than a linear search through all the constants when a new one is
inserted, use a `HashTable` for lookups. The table maps deduplicated
names to the position in the `Chunk`'s `constants` list. If a global
variable name is encountered for a second time, the same index is reused
as the argument for the new bytecode operation.