Allocate storage for globals during compilation
Introduce `GlobalVariables` to represent global storage for a given
interpretation context. Each global name has a constant index associated
with it, and the index is what is written into the bytecode and used at
runtime for global accesses. The compiler establishes the mapping as it
compiles variable name uses. The strings themselves continue to be
deduplicated by the compiler.