Compile error for variable use in own initializer
Instead of a runtime error (it would end up as a precondition failure at
lookup time in `RawStack`).

Local variables now start with an invalid value for `depth` (it is
`nil`), which is updated after the initializer expression is
successfully compiled. If a reference to the same name is found in the
initializer expression, that is resolved to the name in the current
scope, which cannot be, since it doesn't have a value yet.

Restored the 'outer-in-inner-decl.lox' test file to demonstrate this
case (it was basically just recapitulating the functionality test in
'shadowing.lox').

It may be an interesting exercise to try letting the reference resolve
to _another_ variable, though -- so looking for a name in an enclosing
scope, or a global.