Allow variable refs in own initializers
When resolving local references, lookup just skips any that are
unintialized, falling back to earlier declarations and ultimately to a
global. (And therefore, since globals are finally resolved only at
runtime, it will error out then if the lookup fails.)

The 'outer-in-inner-decl.lox' test file has been updated to demonstrate
the new behavior.