Revert "Allow variable refs in own initializers"
I liked this way a lot better; the compiler code is simpler and I think
the behavior in Lox code makes more sense. But it's not the way the book
does it, and I am afraid that's going to make following along more
complicated later. So I will stick with the offical behavior in this
case.

This reverts commit 4b781ca1b0dd0b35b842d0a90b1b9f4e950ea4c1.