Jump directly to addresses instead of by offsets
A jump instruction's argument is now simply the destination index in the
bytecode, rather than an offset relative to the instruction's location.
This obviates the distinction between jump and loop instructions;
instead there are `.jump` and `.jumpLong`. A loop will use one or the
other as appropriate for the operand; a forward jump has to use the long
variant because the destination is not known at the time space in the
bytecode is set aside for the operand.

The `InstructionPointer` accordingly supports updating its `address` to
a particular value directly, which is how the VM interprets these
instructions.