Support 'switch' statements
Introduce a new "match" opcode for checking a switch's case patterns.
The VM interprets this to compare the value and pattern with `==` but
also leaves the value on the stack if they do not match. That way
successive patterns can reuse the value without re-interpreting the
switch's expression.

This is a slightly more complex instruction compared to the "official"
solution of adding an opcode to duplicate the value on the top of the
stack. Eventually I want to try to extend the interpretation of `.match`
to allow classes to customize their match logic.