Support 'break' statements in loops
This introduces another piece of state to the `Compiler` -- the bytecode
locations for breaks need to be recorded as they are encountered, so
that their jumps can be patched at the end of the current loop.

The 'break.lox' test file exposed a bug in how `if` statements pops were
emitted; this is fixed.