Add 'finally' statement, running on scope exit
A series of `finally` statements make a stack -- those that come later
in a scope execute _earlier_ (as with Swift's `defer`). Its body may be
any statement, so single statements are allowed un-braced. This adds
more `Compiler` state; as with `break`s, the locations of `finally`
statements need to be recorded during block compilation and fixed up at
the end.

Add a test file for the statement.

Although Swift, Go, and Zig call this 'defer', I think 'finally' is a
better keyword for it. Ideally the name would also help convey the stack
nature of multiple statements, but I couldn't come up with anything else
that I liked.